+++ src/index.html
@@ -21,8 +21,11 @@
   <button id="chartView">View Productivity Reports</button>
 
   <!-- ChartJS Component -->
-  <div>
-    <canvas id="myChart" width="400" height="400"></canvas>
+  <div class="charts">
+    <canvas id="chartByStatus" width="400" height="400"></canvas>
+  </div>
+  <div class="charts">
+    <canvas id="chartByCategory" width="400" height="400"></canvas>
   </div>
   <!-- end ChartJS -->
