+++ src/styles/_04defaults.scss
@@ -171,3 +171,6 @@ img{
     display: none;
 }
 
+.charts {
+    display: inline-block;
+}
